Исправлен баг с начислением очков main.cpp

+++ main.cpp
@@ -14,7 +14,7 @@ int main()
 		cin >> x >> y;
 		pair<int, int> i = {x + rand() % 11 - 5, y + rand() % 11 - 5};
 		int r = pow(i.first, 2) + pow(i.second, 2);
-		sum += (((int)(5 - sqrt(r)) >> 31) & 1) == 0 ? 5 - sqrt(r) : 0;
+		sum += (((int)(5 - sqrt(r)) >> 31) & 1) == 0 ? 5 - (int)sqrt(r) : 0;
 		cntShot--;
 	}
 	if (sum < 10)
